backend/python/{{cookiecutter.github_repository_name}}/requirements.txt
boto3
watchtower
logstash-formatter
prometheus-client
kafka-python